// src/types/multisig/multisigFunctionNames.ts
export enum MultisigContractFunction {
  SIGN = 'sign',
  UNSIGN = 'unsign',
  DISCARD_ACTION = 'discardAction',
  PERFORM_ACTION = 'performAction',
  PROPOSE_CHANGE_QUORUM = 'proposeChangeQuorum',
  PROPOSE_ADD_BOARD_MEMBER = 'proposeAddBoardMember',
  GET_ALL_BOARD_MEMBERS = 'getAllBoardMembers',
  PROPOSE_ADD_PROPOSER = 'proposeAddProposer',
  PROPOSE_REMOVE_USER = 'proposeRemoveUser',
  PROPOSE_TRANSFER_EXECUTE = 'proposeTransferExecute',
  PROPOSE_ASYNC_CALL = 'proposeAsyncCall',
  PROPOSE_SC_DEPLOY_FROM_SOURCE = 'proposeSCDeployFromSource',
  PROPOSE_SC_UPGRADE_FROM_SOURCE = 'proposeSCUpgradeFromSource',
  ISSUE = 'issue',
  ESDT_TRANSFER = 'ESDTTransfer',
  ESDT_NFT_TRANSFER = 'ESDTNFTTransfer',
  GET_PENDING_ACTION_FULL_INFO = 'getPendingActionFullInfo',
  GET_NUM_BOARD_MEMBERS = 'getNumBoardMembers',
  GET_NUM_PROPOSERS = 'getNumProposers',
  GET_QUORUM = 'getQuorum',
  GET_ACTION_LAST_INDEX = 'getActionLastIndex',
  GET_ACTION_DATA = 'getActionData',
  USER_ROLE = 'userRole',
  GET_ALL_PROPOSERS = 'getAllProposers',
  GET_ACTION_SIGNERS = 'getActionSigners',
  GET_ACTION_VALID_SIGNER_COUNT = 'getActionValidSignerCount',
  GET_ACTION_SIGNER_COUNT = 'getActionSignerCount',
  QUORUM_REACHED = 'quorumReached',
  SIGNED = 'signed',
  REGISTER_MULTISIG_CONTRACT = 'registerMultisigContract',
  REGISTER_MULTISIG_NAME = 'registerMultisigName',
  UNREGISTER_MULTISIG_CONTRACT = 'unregisterMultisigContract',
  GET_MULTISIG_CONTRACTS = 'getMultisigContracts',
  GET_MULTISIG_CONTRACT_NAME = 'getMultisigContractName'
}